# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API документация для e-commerce",
        "description": "Эта документация описывает REST API для интернет-магазина",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Получить корзину пользователя",
                "operationId": "399a5f817b36b0db308a434d07715bad",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cart_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Ноутбук Apple MacBook Pro"
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1999.99
                                                    },
                                                    "total": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 3999.98
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 3999.98
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Добавить товар в корзину",
                "operationId": "af5c3a2c065e9b55aaa49e4cf9b0f9ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Товар добавлен в корзину"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/remove": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Удалить товар из корзины",
                "operationId": "448f8358d222be2f1a431168f05dd7f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Товар удален из корзины"
                    },
                    "404": {
                        "description": "Товар не найден"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/clear": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Очистить корзину пользователя",
                "operationId": "baaa3721c515b660857812df757013a2",
                "responses": {
                    "200": {
                        "description": "Корзина очищена"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}